requirements.txt
onnx
onnxruntime-gpu
torchmetrics
numpy
pandas
tqdm
pillow
pycocotools
timm